refactor(actions): split panel, preferences, and workflow definition files by sub-domain
- Split panelActions.ts (871→15 line barrel + 4 sub-domain files: panelCore, portal, portalTab, portalHelpers)
- Split preferencesActions.ts (863→22 line barrel + 6 sub-domain files: prefsUi, window, keybinding, help, terminalConfig, appConfig)
- Split workflowActions.ts (660→11 line barrel + 3 sub-domain files: workflowHelpers, workflowCreation, workflowUtility)
- Updated helpLaunchAgent.test.ts import to use registerHelpActions directly
- Follows existing terminal (7 files) and worktree (9 files) split convention
